+++ homepage/homepage/components/footer.tsx
@@ -1,45 +1,24 @@
+import { socials } from "@/lib/socials";
 import { GcmpLogo } from "gcmp-design-system/src/app/components/atoms/logos/GcmpLogo";
 import { Footer } from "gcmp-design-system/src/app/components/organisms/Footer";
 
 export function JazzFooter() {
   return (
     <Footer
-      logo={<GcmpLogo monochrome className="w-32" />}
+      logo={<GcmpLogo monochrome className="w-36" />}
       companyName="Garden Computing, Inc."
+      socials={socials}
       sections={[
         {
-          title: "Resources",
-          links: [
-            { href: "/", label: "Toolkit" },
-            { href: "/cloud", label: "Jazz Cloud" },
-            { href: "/docs", label: "Docs" },
-          ],
-        },
-        {
-          title: "Community",
+          title: "About",
           links: [
             {
-              href: "https://github.com/gardencmp/jazz",
-              label: "GitHub",
+              href: "https://garden.co/team",
+              label: "Team",
               newTab: true,
             },
             {
-              href: "https://discord.gg/utDMjHYg42",
-              label: "Discord",
-              newTab: true,
-            },
-            {
-              href: "https://x.com/jazz_tools",
-              label: "X",
-              newTab: true,
-            },
-          ],
-        },
-        {
-          title: "News",
-          links: [
-            {
-              href: "https://gcmp.io/news",
+              href: "https://garden.co/news",
               label: "Blog",
               newTab: true,
             },
@@ -47,6 +26,23 @@ export function JazzFooter() {
               href: "https://github.com/gardencmp/jazz/releases",
               label: "Releases",
               newTab: true,
+            },
+          ],
+        },
+        {
+          title: "Resources",
+          links: [
+            {
+              href: "/docs",
+              label: "Documentation",
+            },
+            {
+              href: "/examples",
+              label: "Examples",
+            },
+            {
+              href: "/showcase",
+              label: "Built with Jazz",
             },
           ],
         },