+++ homepage/homepage/components/home/EarlyAdopterSection.tsx
@@ -1,10 +1,11 @@
 import { Button } from "gcmp-design-system/src/app/components/atoms/Button";
+import { Card } from "gcmp-design-system/src/app/components/atoms/Card";
 import { H2 } from "gcmp-design-system/src/app/components/atoms/Headings";
 import { Prose } from "gcmp-design-system/src/app/components/molecules/Prose";
 
 export function EarlyAdopterSection() {
   return (
-    <div className="border rounded-xl shadow-sm p-4 md:py-16">
+    <Card className="p-4 md:py-16">
       <div className="lg:max-w-3xl md:text-center mx-auto space-y-6">
         <p className="uppercase text-blue tracking-widest text-sm font-medium dark:text-stone-400">
           Become an early adopter
@@ -31,6 +32,6 @@ export function EarlyAdopterSection() {
           </Button>
         </div>
       </div>
-    </div>
+    </Card>
   );
 }