+++ packages/jazz-react-native-media-images/CHANGELOG.md
@@ -1,5 +1,36 @@
 # jazz-browser-media-images
 
+## 0.8.25
+
+### Patch Changes
+
+- jazz-tools@0.8.34
+
+## 0.8.24
+
+### Patch Changes
+
+- Updated dependencies [df42b2b]
+  - jazz-tools@0.8.32
+
+## 0.8.23
+
+### Patch Changes
+
+- jazz-tools@0.8.31
+
+## 0.8.22
+
+### Patch Changes
+
+- jazz-tools@0.8.30
+
+## 0.8.21
+
+### Patch Changes
+
+- jazz-tools@0.8.29
+
 ## 0.8.20
 
 ### Patch Changes